Fix: TestNewItem
